Reorganized tests directory to have more structure
