+++ tests/bidsDemoData/sub-002/ses-1/beh/sub-002_ses-1_task-FaceRecognition_beh.tsv
