+++ tests/bidsDemoData/sub-002/ses-1/eeg/sub-002_ses-1_task-FacePerception_channels.tsv
