+++ tests/bidsDemoData/sub-002/sub-002_scans.json
