+++ tests/bidsDemoData/sub-003/ses-1/eeg/sub-003_ses-1_task-FacePerception_run-1_eeg.json
