+++ tests/bidsDemoData/sub-003/sub-003_scans.json
