+++ tests/bidsDemoData/sub-004/ses-1/beh/sub-004_ses-1_task-FaceRecognition_beh.tsv
