+++ tests/bidsDemoData/sub-004/ses-1/eeg/sub-004_ses-1_coordsystem.json
