+++ tests/bidsDemoData/sub-004/ses-1/eeg/sub-004_ses-1_task-FacePerception_run-1_events.tsv
