+++ tests/bidsDemoData/sub-004/ses-1/sub-004_ses-1_scans.tsv
