+++ tests/otherTestData/HED_score_1.0.0.xml
