+++ tests/otherTests/regexTests.spec.js
