+++ tests/otherTests/string.spec.js
